res_drivers_license: renamed from res_drivers_licenses and added fields
