+++ res_drivers_license/models/res_drivers_license.py
@@ -25,7 +25,7 @@ _logger = logging.getLogger(__name__)
 class ResDriversLicense(models.Model): 
     _name = 'res.drivers_license'
 
-    partner_id = fields.Many2one(comodel_name="res.partner")
+    partner_id = fields.Many2many(comodel_name="res.partner")
     
     name = fields.Char(string='Type', required=True) #A,B etc.
     description = fields.Char(string='Description')
