+++ res_drivers_license/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_drivers_license_manager,res.drivers_license,res_drivers_license.model_res_drivers_license,base.group_partner_manager,1,1,1,1
+access_drivers_license_user,res.drivers_license.user,res_drivers_license.model_res_drivers_license,base.group_user,1,0,0,0
