+++ res_drivers_license/views/res_drivers_license_view.xml
@@ -27,7 +27,7 @@
                 </tree>
             </field>
         </record>
-        <record id="view_drivers_license_filter" model="ir.ui.view">
+        <record id="res_view_drivers_license_filter" model="ir.ui.view">
             <field name="name">res.drivers_license.select</field>
             <field name="model">res.drivers_license</field>
             <field name="arch" type="xml">
@@ -39,12 +39,12 @@
             </field>
         </record>
         <record id="action_drivers_license_form" model="ir.actions.act_window">
-            <field name="name">drivers_license</field>
+            <field name="name">Drivers licenses</field>
             <field name="type">ir.actions.act_window</field>
             <field name="res_model">res.drivers_license</field>
             <field name="view_type">form</field>
             <field name="view_mode">tree,form</field>
-            <field name="view_id" ref="drivers_license.view_drivers_license_tree" />
+            <field name="view_id" ref="res_drivers_license.view_drivers_license_tree" />
             <field name="context">{}</field>
             <field name="search_view_id" ref="res_view_drivers_license_filter" />
         </record>
